requirements.txt
Django>=4.2.0
django-environ>=0.10.0
django-crispy-forms>=2.0
django-filter>=23.2
djangorestframework>=3.14.0
Pillow>=9.5.0
psycopg2-binary>=2.9.6
python-dateutil>=2.8.2
django-simple-history>=3.3.0
django-widget-tweaks>=1.4.12
reportlab>=4.0.4